+++ front/src/pages/trade/order/Order.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useRef, useEffect } from 'react';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
 import { IAskOrderItem, IBidOrderItem, askOrdersAtom, bidOrdersAtom } from '@recoil/stockOrders/index';
-import bidAskPriceAtom from '@src/recoil/bidAskPrice/atom';
+import bidAskPriceAtom from '@recoil/bidAskPrice';
 import formatNumber from '@src/common/utils/formatNumber';
 import AskOrderItem from './AskOrderItem';
 import BidOrderItem from './BidOrderItem';
