+++ front/src/recoil/bidAskPrice/index.ts
@@ -1,3 +1,1 @@
-import atom from './atom';
-
-export default atom;
+export { default } from './atom';
